Switched from MSTest to NUnit (Read Description)
It was discovered that MSTest would not fulfill our needs. Because of this, the project was switched to NUnit.
No functional code was changed, however all MSTest syntax was changed to NUnit syntax.
In addition, the test library was changed to a console application so that the tests could be run in the command line by simply running the exe.
To accomplish the transition, the project had to be remade from the ground up. During this transition, dependencies were cleaned up (they had become a mess in the attempt to transition without remaking the solution). However, as was stated earlier, despite the need to remake the project from the ground up, the functional code was copied exactly from the previous version, so no changes were made there in the transition.